Return proper error code for file exists, note RFC difference
I completely overlooked that there's an actual error code for
if a file already exists. The server will now return that instead
of access violation. I made a notes section in the readme for
implementation details that differ from the RFCs. In this case,
overwriting files is not really supported by RFC 1350 but the
implementation allows it anyway if the cli flag is set.